+++ src/components/customInput/types.ts
@@ -1,19 +1,12 @@
 import { HTMLInputTypeAttribute } from 'react';
 import { UseFormRegister } from 'react-hook-form';
 
+import { ValidationSchema } from '../../utils/validation/types';
+
 import { UserLoginData } from '../../models/models';
 import { UserRegisterData } from '../../models/models';
 
 type CombinedRegisterType = UseFormRegister<UserLoginData | UserRegisterData>;
-
-export type ValidationSchema = {
-  validate?: (value: string) => boolean | string;
-  required: string;
-  pattern?: {
-    value: RegExp;
-    message: string;
-  };
-};
 
 export type TProps = {
   containerClass?: string;
