+++ src/utils/validation/types.ts
@@ -1,0 +1,8 @@
+export type ValidationSchema = {
+  validate?: (value: string) => boolean | string;
+  required: string;
+  pattern?: {
+    value: RegExp;
+    message: string;
+  };
+};
